requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
kagglehub
matplotlib
seaborn
